requirements.txt
google-auth
google-oauth
google-cloud
google-cloud-storage